# api/openapi.yaml
openapi: 3.1.0
info:
  title: Seva API
  description: |
    API specification for the Seva platform.
    This is the single source of truth that generates Go server stubs,
    SvelteKit TypeScript client, and Flutter Dart client.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: engineering@seva.io
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.staging.seva.io
    description: Staging
  - url: https://api.seva.io
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/routes:
    get:
      operationId: listRoutes
      summary: List provider's routes
      tags: [Routes]
      responses:
        "200":
          description: Routes retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createRoute
      summary: Create a new route
      description: Creates a new service route for a provider. Used by crop and land service workers.
      tags: [Routes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteCreate"
      responses:
        "201":
          description: Route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/routes/{id}:
    get:
      operationId: getRoute
      summary: Get route details with stops
      tags: [Routes]
      parameters:
        - $ref: "#/components/parameters/RouteIdPath"
      responses:
        "200":
          description: Route details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/routes/{id}/stops:
    post:
      operationId: addRouteStop
      summary: Add a stop to a route
      tags: [Routes]
      parameters:
        - $ref: "#/components/parameters/RouteIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteStopCreate"
      responses:
        "201":
          description: Stop added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteStop"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/routes/{id}/stops/{stopId}:
    delete:
      operationId: removeRouteStop
      summary: Remove a stop from a route
      tags: [Routes]
      parameters:
        - $ref: "#/components/parameters/RouteIdPath"
        - name: stopId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Stop removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/routes/{id}/optimize:
    post:
      operationId: optimizeRoute
      summary: Optimize route stop order
      description: Reorders stops by geographic proximity for efficient travel.
      tags: [Routes]
      parameters:
        - $ref: "#/components/parameters/RouteIdPath"
      responses:
        "200":
          description: Route optimized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/routes/schedule:
    get:
      operationId: getWeeklySchedule
      summary: Get weekly visit schedule
      description: Returns all scheduled visits for the current week.
      tags: [Routes]
      responses:
        "200":
          description: Weekly schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledVisit"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/routes/request:
    post:
      operationId: requestRouteService
      summary: Request to be added to a route
      description: Customer requests a provider to add them to an existing route.
      tags: [Routes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [route_id, address, latitude, longitude]
              properties:
                route_id:
                  type: string
                address:
                  type: string
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
                tree_count:
                  type: integer
                area_sqft:
                  type: number
                notes:
                  type: string
      responses:
        "201":
          description: Request submitted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Search
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the /auth/verify-otp or /auth/refresh endpoints.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    RouteIdPath:
      name: id
      in: path
      required: true
      description: Route ID
      schema:
        type: string
      example: "rte_01HT2NKFP3R7E3DYGZM2PFWT4F"
...
  responses:
    BadRequest:
      description: Bad request — invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: BAD_REQUEST
            message: Invalid request parameters
            details:
              - field: phone
                message: Phone number must include country code
    Unauthorized:
      description: Unauthorized — missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: Authentication required
            details: null
...
    NotFound:
      description: Not found — the requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: Resource not found
            details: null
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples:
            - "BAD_REQUEST"
        message:
          type: string
          description: Human-readable error message
          examples:
            - "Invalid request parameters"
        details:
          description: Additional error details (validation errors, etc.)
          oneOf:
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            - type: "null"

    # --- Notifications ---
...
    RouteStatus:
      type: string
      enum: [active, paused, completed]
    Route:
      type: object
      required: [id, provider_id, name, postcode_area, interval_days, status, stop_count]
      properties:
        id:
          type: string
        provider_id:
          type: string
        name:
          type: string
        postcode_area:
          type: string
        max_stops:
          type: integer
        stop_count:
          type: integer
        interval_days:
          type: integer
          description: Days between visits (e.g., 30 for monthly)
        status:
          $ref: "#/components/schemas/RouteStatus"
        created_at:
          type: string
          format: date-time
    RouteCreate:
      type: object
      required: [name, postcode_area, interval_days]
      properties:
        name:
          type: string
        postcode_area:
          type: string
        max_stops:
          type: integer
          default: 20
        interval_days:
          type: integer
          minimum: 1
          maximum: 365
    RouteStop:
      type: object
      required: [id, route_id, customer_id, address, latitude, longitude, status]
      properties:
        id:
          type: string
        route_id:
          type: string
        customer_id:
          type: string
        customer_name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        tree_count:
          type: integer
        area_sqft:
          type: number
        last_visited:
          type: string
          format: date-time
        next_visit:
          type: string
          format: date-time
        notes:
          type: string
        status:
          type: string
          enum: [active, paused, removed]
    RouteStopCreate:
      type: object
      required: [customer_id, address, latitude, longitude]
      properties:
        customer_id:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        tree_count:
          type: integer
        area_sqft:
          type: number
        notes:
          type: string
    RouteDetail:
      allOf:
        - $ref: "#/components/schemas/Route"
        - type: object
          properties:
            stops:
              type: array
              items:
                $ref: "#/components/schemas/RouteStop"
    ScheduledVisit:
      type: object
      required: [route_id, route_name, stop_id, customer_name, address, scheduled_date]
      properties:
        route_id:
          type: string
        route_name:
          type: string
        stop_id:
          type: string
        customer_name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        scheduled_date:
          type: string
          format: date
        tree_count:
          type: integer
        notes:
          type: string

    # --- Search Results ---
...